refactor: merge ui and layout folders
