+++ src/app/page.tsx
@@ -1,7 +1,7 @@
 import { EventDto, EventsService, OpenAPI } from '@losol/eventuras';
 
 import { EventGrid } from '@/components/event';
-import { Container, Layout } from '@/components/layout';
+import { Container, Layout } from '@/components/ui';
 import Heading from '@/components/ui/Heading';
 import Text from '@/components/ui/Text';
 import Environment, { EnvironmentVariables } from '@/utils/Environment';
