+++ src/app/user/page.tsx
@@ -1,7 +1,7 @@
 'use client';
 import { useContext } from 'react';
 
-import { Container, Layout } from '@/components/layout';
+import { Container, Layout } from '@/components/ui';
 import Heading from '@/components/ui/Heading';
 import Loading from '@/components/ui/Loading';
 import { UserContext } from '@/context/UserContext';
